+++ src/components/layout/layout.tsx
@@ -9,8 +9,8 @@ import * as React from "react"
 import PropTypes from "prop-types"
 import { useStaticQuery, graphql } from "gatsby"
 
-import Header from "./header"
-import "./layout.css"
+import Header from '../header/header'
+import "./layout.scss"
 
 interface LayoutProps {
   children: React.ReactNode
