+++ src/pages/404.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 
 import Layout from "../components/layout"
-import Seo from "../components/seo"
+import Seo from "../components/seo/seo"
 
 const NotFoundPage = () => (
   <Layout>
