+++ src/pages/index.tsx
@@ -3,7 +3,7 @@ import { Link } from "gatsby"
 import { StaticImage } from "gatsby-plugin-image"
 
 import Layout from "../components/layout"
-import Seo from "../components/seo"
+import Seo from "../components/seo/seo"
 
 const IndexPage = () => (
   <Layout>
